Configure datastax-agent and add OpsCenter node
Just needed to properly tell the datastax-agent where to connect to.
Also needed to bump up the RAM on the nodes a bit to be able to
support the agent along with the Cassandra instance. I'll work to
continue tweaking, see if I can get this back down to 768MB.